requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.12
sse-starlette>=2.1.0
aiofiles>=24.1.0

# HTTP + parsing
requests>=2.32.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Config + data
pyyaml>=6.0.0
ruamel.yaml>=0.18.0
python-dotenv>=1.0.0

# Job sources
apify-client>=1.8.0

# Output generation
playwright>=1.49.0
openpyxl>=3.1.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
